Extract polling timestamp and reaction grouping into concerns
The server-timestamp sync format and the journal reaction grouping each now
live in a dedicated concern. UpdateStreaming shares the journal-id grouping
helper rather than spelling out the GroupedQueries call itself.
